+++ R/Model.R
@@ -26,6 +26,15 @@
 #' * `model`\cr
 #' The stored model object in its original form.
 #'
+#' @section Public fields:
+#'
+#' * `preprocessing_fn`\cr
+#' Default as NULL, this is to store a preprocessing function which will be
+#' used to evaluate the entity data in [Trans] prior to simulating the
+#' transition. A situation where this is useful could be when you want to limit
+#' the use of a [Model] object to the specific group of agents (e.g: age between
+#' `x` and `y`) that was used to estimate the model.
+#'
 #' @section Public Methods:
 #'
 #'  * `get()`\cr
@@ -50,7 +59,8 @@
 #'
 #' simple_prob_model <- Model$new(x = list(yes = 0.95, no = 0.05))
 #'
-#' simple_glm_model <- Model$new(x = stats::glm(factor(sex) ~ age, data = toy_individuals, family = "binomial"))
+#' simple_glm_model <- Model$new(x = stats::glm(factor(sex) ~ age,
+#'                               data = toy_individuals, family = "binomial"))
 #'
 #' # return the original model object
 #' simple_prob_model$model
@@ -76,7 +86,8 @@ Model <-
     classname = "Model",
     inherit = Generic,
     public = list(
-      initialize = function(x) {
+      initialize = function(x, preprocessing_fn = NULL) {
+        self$preprocessing_fn <- preprocessing_fn
         self$set(x)
       },
       get = function() {
@@ -98,7 +109,8 @@ Model <-
       },
       print = function() {
         print(private$.model)
-      }
+      },
+      preprocessing_fn = NULL
     ),
     active = list(
       model = function() {
@@ -119,5 +131,9 @@ Model <-
 #' @export
 #' @rdname Model
 summary.Model <- function(object, ...) {
-  summary(object$model)
+  if (object$class() == "WrappedModel") {
+    summary(object$model$learner.model)
+  } else {
+    summary(object$model)
+  }
 }