+++ NEWS.md
@@ -1,3 +1,17 @@
+# dymiumCore 0.1.6
+
+## NEW FEATURES
+- Added `transition()`, this is a function for simulating state transition inside a microsimulation pipeline. It also comes with `get_transition()` which returns state transition outcomes. See its documentation, `?dymiumCore::transition`, for more detail. 
+- [Model] gains `preprocessing_fn` field, this is for storing a preprocessing function which will be used to evaluate the entity data in [Transition] prior to simulating the transition. A situation where this is useful could be when you want to limit the use of a [Model] object to the specific group of agents (e.g: age between `x` and `y`) that was used to estimate the model.
+- [Trans] supports [Model]'s `preprocessing_fn()`.
+- Minor fix to error msg in [TransitionClassification].
+- [add_models] gains a as_r6model for returning models as [Model] objects.
+- Classification models fitted using the `mlr` package is now supported by `transition`.
+- Added `simulate_choice`, a function for simulating choice selection from the predicted probability obtained from a model object, currently only supports most of the classification models fitted with `caret` and `mlr`, and `glm` of the binomial family.
+
+## INTERNAL CHANGES
+- Renamed `Transition` to `Trans` to make way for `transtion()`.
+
 # dymiumCore 0.1.5
 
 ## NEW FEATURES